LayoutTests/js/array-iterate-backwards-expected.txt

This test checks that iterating a large array backwards works correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS verify(32768) is true
PASS verify(65536) is true
PASS verify(120000) is true
PASS successfullyParsed is true

TEST COMPLETE
